requirements.txt
telebot
psycopg2-binary
ccxt
asyncio
beautifulsoup4